- added support VFO A and B - freq is now unsigned long

+++ avr/transceiver.h
@@ -3,7 +3,7 @@
 int init_geraet(void);
 int tx(void);
 int rx(void);
-int tune(unsigned int freq2tune,unsigned int step2tune);
+int tune(unsigned long freq2tune,unsigned int step2tune);
 int modulation(unsigned int mod);
 int rogerbeep(void);
 void off(void);
